Create Bmaster
Example
Input:
3 3 3
1 1
1 1
1 1
1 1
1 1
1 1
1 1
1
2
3

Output:
4
12
17
Explanation
The initial grid looks as follows:
100
000
000
Here '1' represents a cell with an image and '0' represents a free cell.

After the first move grid is
111
000
000

After the second move grid is
101
111
111

Finally, after the third move we have
011
101
010

So after the first move Maelk will see 3 images, after the second move – 8 images and after the third move – 5 images. Hence the answers for T = 1, 2, 3 are 1 + 3 = 4, 1 + 3 + 8 = 12 and 1 + 3 + 8 + 5 = 17 respectively.